Move the non-working "Create desktop shortcut" stuff to the request framework
