+++ Windows/W32Util/ShellUtil.cpp
@@ -216,4 +216,79 @@ namespace W32Util
 
 		PostMessage(parent_, completeMsg_, 0, 0);
 	}
+
+
+// http://msdn.microsoft.com/en-us/library/aa969393.aspx
+HRESULT CreateLink(LPCWSTR lpszPathObj, LPCWSTR lpszArguments, LPCWSTR lpszPathLink, LPCWSTR lpszDesc) {
+	HRESULT hres;
+	IShellLink *psl = nullptr;
+	hres = CoInitializeEx(NULL, COINIT_MULTITHREADED);
+	if (FAILED(hres))
+		return hres;
+
+	// Get a pointer to the IShellLink interface. It is assumed that CoInitialize
+	// has already been called.
+	hres = CoCreateInstance(CLSID_ShellLink, NULL, CLSCTX_INPROC_SERVER, IID_IShellLink, (LPVOID *)&psl);
+	if (SUCCEEDED(hres) && psl) {
+		IPersistFile *ppf = nullptr;
+
+		// Set the path to the shortcut target and add the description. 
+		psl->SetPath(lpszPathObj);
+		psl->SetArguments(lpszArguments);
+		psl->SetDescription(lpszDesc);
+
+		// Query IShellLink for the IPersistFile interface, used for saving the 
+		// shortcut in persistent storage. 
+		hres = psl->QueryInterface(IID_IPersistFile, (LPVOID *)&ppf);
+
+		if (SUCCEEDED(hres) && ppf) {
+			// Save the link by calling IPersistFile::Save. 
+			hres = ppf->Save(lpszPathLink, TRUE);
+			ppf->Release();
+		}
+		psl->Release();
+	}
+	CoUninitialize();
+
+	return hres;
 }
+
+bool CreateDesktopShortcut(const std::string &argumentPath, std::string gameTitle) {
+	// TODO: not working correctly
+	return false;
+
+	// Get the desktop folder
+	// TODO: Not long path safe.
+	wchar_t *pathbuf = new wchar_t[MAX_PATH + gameTitle.size() + 100];
+	SHGetFolderPath(0, CSIDL_DESKTOPDIRECTORY, NULL, SHGFP_TYPE_CURRENT, pathbuf);
+
+	// Sanitize the game title for banned characters.
+	const char bannedChars[] = "<>:\"/\\|?*";
+	for (size_t i = 0; i < gameTitle.size(); i++) {
+		for (char c : bannedChars) {
+			if (gameTitle[i] == c) {
+				gameTitle[i] = '_';
+				break;
+			}
+		}
+	}
+
+	wcscat(pathbuf, L"\\");
+	wcscat(pathbuf, ConvertUTF8ToWString(gameTitle).c_str());
+
+	std::wstring moduleFilename;
+	size_t sz;
+	do {
+		moduleFilename.resize(moduleFilename.size() + MAX_PATH);
+		// On failure, this will return the same value as passed in, but success will always be one lower.
+		sz = GetModuleFileName(nullptr, &moduleFilename[0], (DWORD)moduleFilename.size());
+	} while (sz >= moduleFilename.size());
+	moduleFilename.resize(sz);
+
+	CreateLink(moduleFilename.c_str(), ConvertUTF8ToWString(argumentPath).c_str(), pathbuf, ConvertUTF8ToWString(gameTitle).c_str());
+
+	delete[] pathbuf;
+	return false;
+}
+
+}  // namespace
