+++ Windows/WindowsHost.h
@@ -31,8 +31,6 @@ public:
 
 	void ToggleDebugConsoleVisibility() override;
 
-	bool CreateDesktopShortcut(std::string argumentPath, std::string title) override;
-
 	void NotifyUserMessage(const std::string &message, float duration = 1.0f, u32 color = 0x00FFFFFF, const char *id = nullptr) override;
 
 private:
